Give priority to fields in later files processed
Previously, we replaced a field if it was empty. This meant that
fields from earlier files had priority/precedence. Now, we replace
a field if the new field is non-empty. This gives precedence to
fields found in later files.

The effect is equivalent to reversing the order of the input files
so there is no technical value to the change. However, I believe
that it is easier to explain and understand after the change.

We can now simply state that "Fields found in later files override
fields from previously processed files."