Add --excludeID and --excludeCountry options
The --excludeID option takes a single DMR ID to exclude or an
inclusive range of DMR IDs to exclude.  Any records matching
the excluded ID or range are omitted from the output.

The --excludeCountry option takes a country name or abbreviation.
Any records with countries matching the country name or abbreviation
are excluded from the output.